Reorganized dataclasses into a single file
