+++ projects/sandbox/train/train/data_structures.py
@@ -2,8 +2,11 @@ from typing import Callable, Optional, Tuple
 
 import numpy as np
 import torch
+from train.transforms.transform import Transform
 
 from ml4gw.dataloading import InMemoryDataset
+from ml4gw.transforms.injection import RandomWaveformInjection
+from ml4gw.utils.slicing import sample_kernels
 
 
 class BBHInMemoryDataset(InMemoryDataset):
@@ -70,4 +73,66 @@ class BBHInMemoryDataset(InMemoryDataset):
 
         if self.preprocessor is not None:
             X, y = self.preprocessor(X, y)
+        return X, y
+
+
+class BBHNetWaveformInjection(RandomWaveformInjection):
+    def forward(self, X: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
+        if not self.training:
+            return X, y
+
+        X, idx, _ = super().forward(X)
+        y[idx] = 1
+        return X, y
+
+
+# TODO: generalize to arbitrary ifos
+class GlitchSampler(Transform):
+    def __init__(
+        self, prob: float, max_offset: int, **glitches: np.ndarray
+    ) -> None:
+        super().__init__()
+        self.glitches = torch.nn.ParameterList()
+        for ifo in glitches.values():
+            param = self.add_parameter(ifo)
+            self.glitches.append(param)
+
+        self.prob = prob
+        self.max_offset = max_offset
+
+    def forward(self, X: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
+        if X.shape[1] < len(self.glitches):
+            raise ValueError(
+                "Can't insert glitches into tensor with {} channels "
+                "using glitches from {} ifos".format(
+                    X.shape[1], len(self.glitches)
+                )
+            )
+
+        # sample batch indices which will be replaced with
+        # a glitch independently from each interferometer
+        masks = torch.rand(size=(len(self.glitches), len(X))) < self.prob
+        for i, ifo in enumerate(self.glitches):
+            mask = masks[i]
+
+            # now sample from our bank of glitches for this
+            # interferometer the number we want to insert
+            N = mask.sum().item()
+            idx = torch.randint(len(ifo), size=(N,))
+
+            # finally sample kernels from the selected glitches.
+            # Add a dummy dimension so that sample_kernels
+            # doesn't think this is a single multi-channel
+            # timeseries, but rather a batch of single
+            # channel timeseries
+            glitches = ifo[idx, None]
+            glitches = sample_kernels(
+                glitches,
+                kernel_size=X.shape[-1],
+                max_center_offset=self.max_offset,
+            )
+
+            # replace the appropriate channel in our
+            # strain data with the sampled glitches
+            X[mask, i] = glitches[:, 0]
         return X, y
